Removed Flutter Blue from web builds.
Added get state to BackEnd instead of always using Flutter Blue.
