+++ lib/lighthouseProvider/adapterState/AdapterState.dart
@@ -1,0 +1,10 @@
+enum BluetoothAdapterState {
+  unknown,
+  unavailable,
+  unauthorized,
+  turningOn,
+  on,
+  turningOff,
+  off,
+  error
+}
