+++ lib/widgets/ScanningMixin.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_blue/flutter_blue.dart';
 import 'package:lighthouse_pm/lighthouseProvider/LighthouseProvider.dart';
 import 'package:lighthouse_pm/permissionsHelper/BLEPermissionsHelper.dart';
 import 'package:lighthouse_pm/platformSpecific/shared/LocalPlatform.dart';
@@ -13,7 +12,7 @@ abstract class ScanningMixin {
   Future<bool> _onWillPop() async {
     // A little workaround for issue https://github.com/pauldemarco/flutter_blue/issues/649
     if (LocalPlatform.isAndroid) {
-      if (await FlutterBlue.instance.isScanning.first ||
+      if (await LighthouseProvider.instance.isScanning.first ||
           await _hasConnectedDevices()) {
         await LighthouseProvider.instance.cleanUp();
         await Future.delayed(Duration(milliseconds: 100));
